src/holopy/core/hilbert.py
"""
Hilbert space management module implementing holographic principle constraints.
"""
from pathlib import Path
from typing import Dict, Optional, Tuple, Any
import numpy as np
from ..utils.persistence import StatePersistence, CompressionMethod
from ..config.constants import (
    INFORMATION_GENERATION_RATE,
    PLANCK_CONSTANT,
    CRITICAL_THRESHOLD
)
import logging

logger = logging.getLogger(__name__)

class HilbertSpace:
    """Manages quantum state operations in holographic Hilbert space."""
    
    def __init__(
        self,
        dimension: int,
        boundary_radius: float,
        storage_path: Optional[Path] = None
    ):
        """
        Initialize Hilbert space with holographic constraints.
        
        Args:
            dimension: Dimension of the Hilbert space
            boundary_radius: Radius of the holographic boundary
            storage_path: Optional path for state persistence
        """
        self.dimension = dimension
        self.boundary_radius = boundary_radius
        self.max_information = 4 * np.pi * (boundary_radius**2) / (4 * PLANCK_CONSTANT)
        
        # Initialize basis states
        self.basis_states = self._create_holographic_basis()
        
        # Initialize persistence system
        if storage_path is None:
            storage_path = Path.home() / ".holopy" / "states"
        self.persistence = StatePersistence(
            storage_path,
            compression_method=CompressionMethod.BLOSC
        )
        
        logger.info(
            f"Initialized HilbertSpace(d={dimension}, r={boundary_radius}) "
            f"with max_information={self.max_information:.2e}"
        )

    def save(
        self,
        state: np.ndarray,
        metadata: Dict[str, Any],
        timestamp: float,
        is_checkpoint: bool = False
    ) -> Path:
        """Save quantum state with holographic constraints."""
        # Project state before saving
        projected_state = self.project_state(state)
        
        # Validate entropy bound
        if self.calculate_entropy(projected_state) > self.max_information:
            logger.warning("State exceeds holographic entropy bound - applying projection")
            projected_state = self._enforce_entropy_bound(projected_state)
        
        return self.persistence.save_state(
            projected_state,
            metadata,
            timestamp,
            is_checkpoint
        )

    def load(
        self,
        version_id: Optional[str] = None,
        timestamp: Optional[float] = None,
        latest_checkpoint: bool = False
    ) -> Tuple[Optional[np.ndarray], Optional[Dict]]:
        """Load quantum state with holographic validation."""
        state, metadata = self.persistence.load_state(
            version_id=version_id,
            timestamp=timestamp,
            latest_checkpoint=latest_checkpoint
        )
        
        if state is not None:
            # Ensure loaded state satisfies holographic constraints
            state = self.project_state(state)
            state = self._enforce_entropy_bound(state)
            
        return state, metadata

    def project_state(self, state: np.ndarray) -> np.ndarray:
        """Project quantum state onto holographic basis."""
        # Expand in basis
        coefficients = np.array([np.vdot(basis, state) for basis in self.basis_states])
        
        # Apply holographic cutoff
        cutoff = np.exp(-INFORMATION_GENERATION_RATE * np.arange(self.dimension))
        coefficients *= cutoff
        
        # Reconstruct state
        projected = np.zeros_like(state)
        for coeff, basis in zip(coefficients, self.basis_states):
            projected += coeff * basis
            
        return projected / np.sqrt(np.vdot(projected, projected))

    def calculate_entropy(self, state: np.ndarray) -> float:
        """Calculate von Neumann entropy of state."""
        density_matrix = np.outer(state, state.conj())
        eigenvalues = np.linalg.eigvalsh(density_matrix)
        entropy = 0.0
        for eig in eigenvalues:
            if eig > CRITICAL_THRESHOLD:
                entropy -= eig * np.log2(eig)
        return entropy

    def _create_holographic_basis(self) -> np.ndarray:
        """Create holographically constrained basis states."""
        basis = np.zeros((self.dimension, self.dimension), dtype=np.complex128)
        
        for n in range(self.dimension):
            # Create nth eigenstate
            state = np.zeros(self.dimension, dtype=np.complex128)
            state[n] = 1.0
            
            # Apply holographic constraints
            k = 2 * np.pi * n / (self.boundary_radius * self.dimension)
            state *= np.exp(-INFORMATION_GENERATION_RATE * abs(k))
            
            basis[n] = state / np.sqrt(np.vdot(state, state))
            
        return basis

    def _enforce_entropy_bound(self, state: np.ndarray) -> np.ndarray:
        """Enforce holographic entropy bound on state."""
        current_entropy = self.calculate_entropy(state)
        if current_entropy <= self.max_information:
            return state
            
        # Apply progressive projection until bound is satisfied
        projected = state.copy()
        scale_factor = 0.9
        
        while self.calculate_entropy(projected) > self.max_information:
            projected = self.project_state(projected)
            projected *= scale_factor
            scale_factor *= 0.9
            
        return projected / np.sqrt(np.vdot(projected, projected)) 